Add OpenTripCommand, DeleteTripCommand
