+++ src/main/java/seedu/traveltrio/command/OpenTripCommand.java
@@ -1,6 +1,7 @@
-package seedu.duke.command;
-import seedu.duke.model.Trip;
-import seedu.duke.model.TripList;
+package seedu.traveltrio.command;
+
+import seedu.traveltrio.model.Trip;
+import seedu.traveltrio.model.TripList;
 
 public class OpenTripCommand extends TripCommand {
     private final int index;
@@ -17,7 +18,7 @@ public class OpenTripCommand extends TripCommand {
         }
         Trip tripToOpen = tripList.get(index);
         tripToOpen.setOpen(true);
-        return "Opened trip: " + tripToOpen.getDestination();
+        return "Opened trip: " + tripToOpen.toString();
     }
     
 }
